updated config and server
