--- client/src/components/App.js
@@ -1,17 +1,0 @@
-import React from 'react';
-
-import Title from './Title';
-import Profile from './Profile';
-import Scoreboard from './Scoreboard';
-
-const App = () => {
-  <div>
-    {/* <Title /> */}
-    {/* <Scoreboard /> */}
-    {/* <PeachImage /> */}
-    {/* <Comic /> */}
-    {/* <Profile /> */}
-  </div>;
-};
-
-export default App;
