+++ package.json
@@ -11,21 +11,26 @@
   "author": "Angela",
   "license": "",
   "devDependencies": {
-    "babel-core": "^6.23.1",
+    "babel-cli": "^6.26.0",
+    "babel-core": "^6.26.0",
     "babel-eslint": "^8.2.2",
-    "babel-loader": "^6.4.1",
-    "babel-preset-env": "^1.6.1",
+    "babel-loader": "^8.0.0-beta.2",
     "babel-preset-es2015": "^6.22.0",
     "babel-preset-react": "^6.24.1",
     "css-loader": "^0.28.7",
     "eslint": "^4.18.1",
-    "file-loader": "^1.1.6",
-    "image-webpack-loader": "^3.4.2",
-    "webpack": "^4.1.1"
+    "file-loader": "^1.1.11",
+    "image-webpack-loader": "^3.6.0",
+    "webpack": "^4.1.1",
+    "webpack-cli": "^2.0.12"
   },
   "dependencies": {
     "@angular/cli": "^1.6.3",
+    "@babel/core": "^7.0.0-beta.42",
+    "@babel/preset-env": "^7.0.0-beta.42",
+    "@babel/preset-react": "^7.0.0-beta.42",
     "angular": "^1.6.7",
+    "babel-preset-env": "^1.6.1",
     "body-parser": "^1.17.2",
     "css-modules-require-hook": "^4.2.2",
     "express": "^4.15.0",
@@ -34,7 +39,6 @@
     "mysql": "^2.13.0",
     "node-sass": "^4.7.2",
     "nodemon": "^1.17.1",
-    "prop-types": "^15.6.1",
     "react": "^15.4.2",
     "react-dom": "^15.4.2",
     "react-redux": "^5.0.7",
@@ -45,8 +49,7 @@
     "redux-thunk": "^2.2.0",
     "style-loader": "^0.19.1",
     "url-loader": "^0.6.2",
-    "vis": "^4.21.0",
-    "webpack-cli": "^2.0.12"
+    "vis": "^4.21.0"
   },
   "compilerOptions": {
     "experimentalDecorators": true,
