+++ server/index.js
@@ -6,7 +6,7 @@ var items = require('../database-mysql');
 
 var app = express();
 
-app.use(express.static(__dirname + '/../src/dist'));
+app.use(express.static(__dirname + '/../client/dist'));
 
 app.listen(8000, function() {
   console.log('listening on port 8000!');
